% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods2.R
\name{loglM3}
\alias{loglM3}
\title{Likelihood for M3}
\usage{
loglM3(pars, dat)
}
\description{
This model is similar to M2 but allows the inclusions of samples that only
have overall expression and genotypes.
}
