% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods3.R
\name{loglM5}
\alias{loglM5}
\title{Model M5}
\usage{
loglM5(pars, dat, finiteonly = TRUE)
}
\description{
This is the same model as M4 but with the assumption that bxi = 0. It is used
to test besc = 0 under the assumption that there is no effect on Xi. In a
broader sense, it should find any QTL that affects expression by modifying
escape (chance of escape or amount of escape).
}
